Adding FPS Meter and using smart pointers.

+++ src/lib/Point.h
@@ -1,12 +1,15 @@
 #ifndef SAX_POINT_H
 #define SAX_POINT_H
 
+#include <memory>
+#include <iostream>
+
 namespace sax {
 	class Point {
 		public:
-			Point( int x = 0, int y = 0 );
-			int x;
-			int y;
+			Point( double x = 0, double y = 0 );
+			double x;
+			double y;
 	};
 }
